snippets: add a snippet to hash objects for Git
It's often useful when porting code from using SHA-1 to SHA-256 to hash
a Git object in both algorithms.  This can be helpful to verify that the
existing hash is correct and get the new hash at the same time.  To make
this easier, add a snippet that hashes a file both ways and prints the
values on separate lines for easy scripting.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>