tmux: add a theme for tmux
Import a slightly stripped-down version of the Nord theme for tmux.
This is basically the entire tmux theme, just without the environment
variables and JavaScript.

Since we're using Source Code Pro, which is a patched font, use the
arrow characters from the private use area.

Be sure to include copyright and license information.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>