dct-jump: add a data accessor
We currently allow additional files to be imported but lack a good way
to share data across the multiple imported files.  While we could use
some sort of global variable, this is ugly and error-prone, so instead,
create a data accessor that we can use to share data across all
configuration files.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>