playbooks/bootstrap: disable `/tmp` as tmpfs
The author's workflow creates a decent amount of temporary work in
places under `/tmp` and the machine on which they're working may not be
rebooted frequently.  As a result, it's very possible to run out of
space in `/tmp` if it's a tmpfs, since there's a limitation on memory.

In addition, because most development systems these days have fast NVME
SSDs, the extra speed of using `/tmp` as a tmpfs is not necessary and
isn't required for good performance.

Because this causes practical problems, tell systemd not to do that.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>