playbooks: set up fonts where available
On a graphical development system, the author uses Source Code Pro,
which is only shipped in Debian in the `texlive-fonts-extra` package.
As a result, it's important to install that package in such an
environment and configure its use, but also not to fail in environments
which are not graphical.

Create the appropriate configuration for these files, but don't fail if
the fonts are missing.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>