zsh: prefer nex over ex
On Debian and Ubuntu, there is an nvi package which ships nex.  This
version of ex works well without colours or other terminal features,
which is notably not the case for Vim (which attempts to clear the
screen with -e and works passably with -es) and Neovim (which also
clears the screen with -e and does not allow use of -es with a
terminal).

Since ironically this is a better, more featureful ex implementation,
let's be sure to prefer it in our editor configuration when we need a
simple editor.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>