playbooks/bootstrap: install nvi
nvi provides nex, which is the ex implementation which works most
gracefully without a terminal or with a dumb terminal, since it does not
attempt to clear the screen or use colours.  Install it by default on
Debian and Ubuntu.

Note that we do not install it on FreeBSD, since nvi is the default vi
implementation there and thus ex is the right implementation by default.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>