dct-mtree: switch order of src and dest operands
There are two types of lines in a file for dct-mtree: those with the
dest entry and those without.  Those without use the same path for the
source and destination, which practically means that they're designed to
install based on the destination location (that is, the one in one's
home directory), and they usually start with a dot as a consequence.

Those with the dest argument specify the location in the source tree,
which as a practical matter don't have a dot.  This leads to a confusing
mix of items in the mtree files, and it would be better to be
consistent.

To fix this, by default, make the file entry in the mtree the
destination and provide the source argument as a `src` key.  That means
that all of our files will contain the destination location as the first
element and will practically all start with a dot, leading to an
easier-to-read file.

Update both the shell and Ruby version of the mtree code and all of the
mtree files.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>