mtree: use ASCII-8BIT regexes instead of an ASCII-8BIT encoding
We need to handle strings and regexes which handle arbitrary byte
sequences because we run on Unix, which allows those in paths.
Traditionally, we forced the encoding of the Ruby version of our script
to ASCII-8BIT so that our regular expression matching didn't encounter a
mismatch between our strings and the regular expression.

However, that's kind of ugly, and we'd like our source code to be in
UTF-8 instead and restrict our regular expressions to ASCII-8BIT only.
Let's do this by removing the encoding declaration and using the `n`
modifier on the regexes, which forces them into ASCII-8BIT like we want.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>