zshenv: make editor selection table easier to read
Right now, the code in the `bmc_editor` function is hard to read because
there are a large number of possible parameters to consider and they're
listed positionally.  To make this easier to read, expand the text such
that each item has a label, which makes it clearer.

For instance, we can now see that for the graphical options, the
`simple` and `detach` parameters must be identical, while for the
terminal options, `detach` must always be false.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>