ssh/config: send `GIT_PROTOCOL` to remote side
By default, Debian versions of OpenSSH send `LANG` and those variables
starting with `LC_`, which are for setting the locale.  Set these in
case we're on a non-Debian system.  In addition, send `GIT_PROTOCOL`,
which allows Git to negotiate protocol v2 and without which we'd be
stuck on protocol v0.

Note that Git itself tries to send `GIT_PROTOCOL` when appropriately
configured, but do so explicitly here just in case.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>