zlogout: only clear screen on real TTY
When we're using a real TTY, such as a Linux VT, and not a
pseudoterminal, it's helpful to clear the screen on logout to prevent
others from seeing what we've been doing.  However, when running in a
pseudoterminal, we're probably in a terminal emulator or a terminal
multiplexor and it can be useful to scroll back through the history to
pull relevant information out.  In such a case, it's very unlikely that
another user will see our output.

To make picking things from the scrollback easier, only clear the screen
if the TTY is not a pseudoterminal.  We assume that pseudoterminals are
using the Unix 98 / SysV names of `/dev/pts/*` which is available on
Linux and FreeBSD.  Should that not be the case, we will simply clear
the screen as before, which, while potentially annoying, is not harmful.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>