zsh: set up `XDG_STATE_DIR` by default
Create the directory `~/.local/state` by default and set
`XDG_STATE_DIR` if it's set.  Also, if `~/mnt/remote/state` exists and
is set, then use that in preference to `~/.local/state`, since
`~/mnt/remote` is where we mount remote data using Lawn and directories
under it will only exist if that path is mounted.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>