zsh: inherit components from PATH on startup
Right now, we always overwrite the components of the path with our
hard-coded path.  However, it can be useful to inherit some path
components from the environment, such as if one wants to add an extra
version of a compilation toolchain.

Instead, let's sort our hard-coded paths into four buckets: personal
paths, which contain locally-installed and personal binaries; software
paths, which include compiler toolchains; local paths, which include the
local system paths; and core paths, which include the default system
path components managed by the package manager.

Always place the personal paths first, since we want our scripts not to
be overridden by others.  Then, walk the path that we've inherited, and
if we find a path we haven't explicitly handled, insert it; otherwise,
if we have a path in one of our groups, insert that entire group (and,
if necessary, all preceding groups not previously inserted) in our
preferred order.  This preserves our desired order but allows some
additional paths to be inserted in an appropriate place.

If, for some reason, we don't insert all the paths (which would mean
that our inherited path didn't have the code paths at all, which is
unlikely), insert them all at the end.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>