git: always use the XDG config path
We have two possible cases to consider when installing a git config
file: the template case and the fallback (non-template) case.  In the
template case, we want the standard file, except that we also want to
add some templated options.  In the fallback case, we only can provide
the standard file.

Traditionally, we've always installed the standard file in `.gitconfig`
and then optionally installed the templated config in
`.config/git/config`.  However, we'd like to eliminate the non-XDG path,
so we'll only install the `.config/git/config` path.

Preserve the standard file in its location and install it with a special
set of mtree rules if we're not in template mode.  Otherwise, just
extend the template to include the standard file at the end, which means
we now only need one file: the XDG one.

Be sure to add an appropriate dependency on the source file so that the
generated file is regenerated if the source file changes.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>