git: use four spaces for gitconfig
In our templated file, we use four spaces since we use four spaces
everywhere else where there's no other language-specific default.  The
original gitconfig used tabs because long ago so did we, and even though
we switched to four spaces, the gitconfig was not updated.

To keep the output nice and consistent between the templated and standard
config files, re-indent the file and update the Vim modeline to use four
spaces.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>