zshenv: don't remove elements from fpath
`fpath` is the variable holding the function paths used by zsh.  Our
recent refactor duplicated the code which sliced off the first two
elements from this array, which caused a long-standing problem to rear
its head.

The problem is that we continue to slice off elements when we're in a
subshell, so since we start our X11 session in zsh; spawn an
Alacritty window, which starts zsh; and then start tmux, which starts
several more zsh versions, we end up slicing off multiple elements each
time, eventually causing us to lack important functions shipped by zsh.

For instance, we can end up with the following message, which indicates
that we've remove the completion code from being loaded:

    setup_completion:143: command not found: compdef

The obvious question is why we do this, which is easily answered when we
realize what these elements are: /usr/local/share/zsh/site-functions and
/usr/share/zsh/vendor-functions.  It's clear that there was unwanted
system or local changes that overrode or broke the default behaviour.

In any event, those changes are no longer needed, since we're no longer
in that situation, and it would be better to solve them differently by
removing the specific offending elements if necessary should that once
again be a problem.  To fix our broken completion, simply stop slicing
elements off of the `fpath` variable and just append our elements, which
will be de-duplicated thanks to `typeset -U`.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>