zshenv: remove appending of previous path
Our recently updated path handling code should include all the
components from the previous path, just in a different order than
before.  Since we've marked the `PATH` variable with `typeset -U`,
duplicates are removed and we're not really doing anything at all here,
so remove this appending.

Similarly, since `prevpath` is now a local variable, we don't need to
unset it, so don't do so.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>