playbooks: set up proper /tmp cleaning
Because we use `/tmp` very frequently and don't reboot often, we disable
a tmpfs on `/tmp`.  However, we still want `/tmp` cleaned up regularly,
so set up a `tmpfiles.d` configuration to do so on boot.  In addition,
clean `/tmp/user` as well, since libpam-tmpdir configures that not to be
cleaned by default (it uses `d` instead of `D` in its config) and we
want it to be tidy.

Ignore errors since `/etc/tmpfiles.d` won't exist if systemd is not
installed, such as in a container or on FreeBSD.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>