playbooks: always install PPAs on Ubuntu
We pretty much always want to use the Git PPA when it's available, since
it gives us newer versions of Git.  The Neovim PPA currently has
versions newer than those in 22.04, but not newer than those in 24.04.
It doesn't really matter, though, since we'll always install the latest
version, so install it regardless.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>