playbooks: install other useful tools on desktop
On a desktop or laptop, we'll want to do a variety of things, including
run and build containers, play music, read mail, open URLs in mail, use
the system secret store, use the clipboard, and run a large variety of
tmux windows.  Install packages that make this possible when we're
running in a graphical environment, which will be a desktop or laptop.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>