Bump vim-fugitive for reftable support
vim-fugitive has an open pull request for reftable support that has not
yet been merged.  Update the branch to this version until a patched
version has been updated.

Note that we take advantage of GitHub exposing objects through all forks
in a network and therefore do not need to update `.gitmodules`.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>