zshenv: refactor path setup code
We have a lot of duplication in our path setup code.  However, we can
take advantage of the fact that zsh uses dynamic scoping to push some of
this setup code into functions and reduce the duplication.  Be sure to
undefine these functions at the end of the scope to avoid namespace
pollution.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>