Add support for dconf data
Unfortunately, MATE and many related apps don't use flat files for their
data and instead use the dconf format.  However, there are command-line
tools to process this format and dump and load it.

Add a config file for MATE's marco window manager in the format produced
by the dump command and load it if the `dconf-cli` package is installed.
Install this package as part of a bootstrap for a graphical environment.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>