playbooks: only use boolean conditions in `when`
Ansible in Debian unstable becomes very unhappy and refuses to work if
all of the arguments in a condition are not boolean.  In this case, we
just want to determine if we've set the list of packages to null because
we want to skip installing packages, so check for that case explicitly.
This keeps Ansible happy and also accomplishes our goal.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>