vimrc: unset `formatprg` for Go
Some newer version of Vim and Neovim have a Go filetype file that sets
`formatprg` to `gofmt`.  Unfortunately, this means that `gq` does not
wrap and format comments anymore since gofmt does not do that.

Since this is annoying and we have ALE to run gofmt as a fixer anyway,
unset `formatprg` for Go files.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>