unsubscribe: decode MIME encoded-words
MIME specifies a way to encode international text in a header using
encoded-words, which are basically small snippets of quoted-printable
or base64 encoding in a header.  Many senders of email encode their
unsubscription information in this way, so to make use of it, we need to
decode the data.  (A cynic might say that this is to make it more
difficult to unsubscribe, which is probably true.)

The quoted-printable encoding is the same as MIME quoted-printable
except that the underscore always represents a space, and the base64
encoding is identical.  Switch to Ruby for processing the data since it
contains built-in functionality for these encodings in its pack and
unpack code.

Be relatively liberal in the encodings we allow.  Note that we're trying
to accept an HTTP or HTTPS URL, which must necessarily be encoded in
US-ASCII, so we don't care too much what ASCII-compatible encoding we
get.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>