gitconfig: add refresh alias
It's common to want to rebase a branch on an updated version of the main
branch.  However, if the current branch is very out of date, changing to
the main branch, pulling, switching back, and then rebasing tends to
cause lots of files to be changed, resulting in make or other tools
needing to rebuild far too much data.

Instead, let's add an alias, `git refresh`, that simply updates a given
branch from its default remote.  If it's the current branch, then update
the working tree; otherwise, leave it unmodified.  If there's a plus
prefixed to the branch name, then accept a forced update.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>