gitconfig: make refresh alias work better with local branch
There are three common cases of branches that we might want to refresh.
First, there is `HEAD` to specify the current branch.  To allow that and
make it work properly, resolve it to a branch name so that we can look
the branch name up as normal.

Second, there is the current branch.  Using `git fetch --update-head-ok`
is not only discouraged by the Git documentation, it also doesn't update
the index, so if the working tree changes, you're stuck with staged
contents in the index.  Detect this case and use `git pull` instead.

Third is a branch that is not checked out.  Continue to use `git fetch`
in this case, which works fine.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>