mutt: sort thread groups by last date received
When reading a high-volume mailing list such as the Git list, we want to
keep up on what threads are new.  Adjust mutt's configuration so that
the most recent thread groups are those with the most recent messages.
That means that our mailbox will contain all of the interesting threads
at the bottom.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>