playbooks: install OpenSSH
OpenSSH is usually installed by default on most systems, but we want to
make sure it's installed.  If there's a specific package to support the
GSSAPI version, make sure that's installed, too, since we want Kerberos
support.  Ignore if it's not available, such as on older systems, since
the default package will then support GSSAPI.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>