playbooks/bootstrap: enable DNSSEC in local resolver
We have configuration in OpenSSH that uses `VerifyHostKeyDNS`.  This
option requires that the local resolver handle DNSSEC and return
authenticated responses.

In order for DNSSEC to work, we need EDNS0, which allows the larger
responses that are common with DNSSEC, as well as to trust the AD bit
from the recursive resolver, so be sure to set those in
`/etc/resolv.conf` if necessary.

Note that FreeBSD has neither of these options for its resolver, so we
do not set them there.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>