playbooks: deploy private dotfiles by default
We have a set of private dotfiles for scripts and other things that
cannot or should not be distributed publicly.  By default, deploy these,
but allow skipping them as well, and do so in the tests where we will
not have credentials.

Note that we do this in a second stage after the main dotfiles because
we need the SSH configuration that sets `VerifyHostKeyDNS` properly so
we don't have to hard-code the SSH keys.  If private dotfiles are
disabled, then there is no harm from installing the dotfiles a second
time.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>