dct-jump: make editing more generic
Right now, there's a very big dependency on zsh and our shell config in
the edit code.  To make this more generic, accept an optional editor
configuration option in the config file that we can customize and make
the edit operation more generic.

In addition, fix the bug we had where we did not pass the first shell
argument properly, resulting in us needing to use `$0`.  Now that we
pass this argument, we can call things just as Git does, which is the
way `EDITOR` and `VISUAL` are intended to work.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>