dct-jump: fix commands with non-word characters in them
We need to be sure to turn our escaped pattern into a regexp when
matching because otherwise it remains a string and Ruby will then
mismatch it, instead looking for the backslash from the escape.  Do
this.

Be sure to also anchor the regexp at the beginning and end or we will
match the simple `"hello"` pattern instead of the `"hello", /(.+)/`
pattern when we do `hello Toronto`.

Add a test using a dash to make sure this is handled correctly.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>