dct-jump: don't match commands when no arguments given
Currently, we initialize the matches to an empty array.  If there are no
arguments, then we never do any pattern matching and simply return the
matches, which is an empty array and is truthy.  We then invoke that
rule even though there is no good reason to do so.

Fix this by explicitly returning nil to signal no match if there are no
arguments.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>