playbooks/bootstrap: adjust DNSSEC enabling script
Our existing script to enable DNSSEC in `/etc/resolv.conf` has a couple
of problems that prevent it from working properly in all cases.

First, we forgot to include `/etc/resolv.conf` as the file to operate
on, so the script operates on stdin and hangs forever.  Fix this by
including the proper script to modify.

Second, in a GitHub Codespace, `/etc/resolv.conf` is mounted into the
container from outside and cannot be unlinked or renamed over.  To make
this work, open `/etc/resolv.conf` read-write, buffer the desired output,
rewind, truncated, and write the expected output in-place instead of
using the `-i` argument.

Finally, since we're needing to seek, use Ruby, which we will have just
installed above (and therefore can be confident exists), since seeking
in Perl requires importing additional modules to get the correct
constants.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>