spec: clean up temporary directory
Failing to clean up our temporary directory can add quite a bit of junk
to `$TMPDIR`, which can in some cases run the system out of temporary
space.  To be good stewards of system resources and keep things tidy,
add a remover to the spec helper to clean the temporary directory up.

Be sure to duplicate the variable because we will not invoke the
finalizer (and therefore will not clean things up) if we pass the
instance variable.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>