Xsession: template
Right now, there's a set of different desktops that we choose in various
fallback order.  Instead, let's use our configuration file so that we
can always choose the right desktop environment.

In a graphical environment, we will always have Ruby installed from the
distro, so we do not need to worry about templating being available.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>