firefox: create intermediate directories
If we have a Firefox profile, then the directories must already exist,
but mtree wants us to specify all parent directories before specifying
files or subdirectories, so to keep our functionality working with
mtree(1), specify the parent directories.

To avoid errors made by omitting the second `>` before `$@`, always use
it and simply truncate the file up front.  Note that a heredoc is not
handled properly by make (since it wants backslashes as continuation
lines but then emits the backslashes into the text), so we use several
echo statements for tidiness.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>